% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-historical_query_run_time.R
\docType{data}
\name{historical_query_run_time}
\alias{historical_query_run_time}
\title{historical query run time}
\format{
A single value indicating the last time these data were queried from Macebase2:
\describe{
  \item{historical_query_run_time}{year-month-day HH:MM:SS (numeric)}
}

#' @source generated by DB
}
\usage{
historical_query_run_time
}
\description{
historical query run time
}
\keyword{datasets}
