% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pswintergoa.R
\docType{data}
\name{PreSelectivityWinterGOA}
\alias{PreSelectivityWinterGOA}
\title{Winter pre-selectivity-corrected data}
\format{
2 data frames
}
\usage{
data(PreSelectivityWinterGOA)
}
\description{
The winter goa dataset including Shumagins 2001-2008 and Shelikof 1995-2007,
and the ages for Shelikof, organized by interval and length or age.

\itemize{
  \item Interval. Original interval number, not unique in older surveys
  \item Length. Length of fish (cm)
  \item Transect. Original transect designation; NULL transects were removed from these data
  \item Description. The original descriptor in old MB
  \item Numbers. Numbers of fish in that edsu
  \item Biomasskg. Biomass (kg) of fish in that edsu
  \item START_TIME. Date and Time (UTC) at the start of the interval
  \item Region. Region in which the survey was performed, used st_join(st_within)
         from the sf library and updated geographic regions stored in:
       'G://ARCVIEW//exported_shapefiles//winter_survey_regions_BarnabasChiniak_NAD1983'
       to determine survey area
  \item Latitude. Latitude at start of edsu; older surveys don't have an end lat/long
  \item Longitude. Longitude at start of edsu; older surveys don't have an end lat/long

}
}
