% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-query_run_time.R
\docType{data}
\name{query_run_time}
\alias{query_run_time}
\title{query run time}
\format{
A single value indicating the last time these data were queried from Macebase2:
\describe{
  \item{query_run_time}{year-month-day HH:MM:SS (numeric)}
}

#' @source generated by DB
}
\usage{
query_run_time
}
\description{
query run time
}
\keyword{datasets}
